util.Log.java & servlet.WMServlet.java
What I made (Part 1):
In case of a EXCPT or a WARN message,
a SevereError string is filled.
there is also a log.getSevereError() which can give back the
error.
there is also a log.setSevereError() which empties
the error.
What I made (Part 2):
A post method, that will be called after the handle() method is finished.
It can be overruled by a post() method in a Servlet.
There is the spot to ask for the log.getSevereError() and do whatever
you want with it. (In my case, I would trigger some email to the development team).
I hope that both parts are useable.