More minor performance tweaks.  Using my little rawbytes test, it has jumped from 335 requests/second to about 352 requests/second.  That's it.  I give up.
At the risk of duplicating code, I unrolled the other write() method in BBOS
and gave each FastWriter knowledge of the Pool it belongs to.  This avoids an
call on the stack for every BBOS.write() and a hashtable.get() on every FW.close().

I played around with a template giving the FW a hint as to how big its buffer
should be, but it made no measureable difference in performance, so I decided
against.  One less thing to keep up with.

Memory consumption by FW is still an issue.