Updated the README file for the contrib section.
Here's how this works:

the anonymous user "wm" has *full* read/write access to the contrib
repository via the standard CVSROOT (:pserver:wm@cvs.webmacro.org:/root).

To access the contrib repository:

   cvs login  (password="webmacro")
   cvs checkout wmcontrib

As long as commits are done via the "wmcontrib" module, we will not receive
these silly email messages.

However, the anonymous user can also read/write the "contrib/" directory of
the "webmacro" module.  If commits are done via that module/directory, we
will be notified via email.

Lane, would you mind testing this a few times as both yourself and as "wm", and
then let the list know how this works?  I'll manage the the server, but the
contrib/ stuff is all yours!  :)