fixed bug #37, which is that response.setContentType() (b/c of #set $Response.ContentType="text/html") was being called after response.getOutputStream().
Apparently, this causes problems with JSDK 2.3 and is in violation of the spec.

This required writing the template to a fastwriter that doesn't have an output stream, forcing fastwriter to buffer internally, then after template expansion, writing the fast wrtier to the output stream.

now that it does this we can also better adjust the exception handling, but I didn't do that.  Maybe a job for Keats or Brian.