don't re-allocate a default sized byte[] array on reset(), just move the pointer back to the beginning.
currently we are only ever calling this just before we throw the ByteBufferOutputStream away so it is really pointless.  If we were going to re-use it, then it would still be moderately pointless as normal methods would not expose the legacy data left from the previous usage, the only argument would be in holding onto excessively large arrays from single large templates followed by multiple small ones, but we can cross that bridge when (if) we come to it.